+++ unit_tests/reduct/auth/token_repository_test.cc
@@ -12,7 +12,7 @@ using reduct::core::Error;
 static std::unique_ptr<ITokenRepository> MakeRepo(std::filesystem::path path = BuildTmpDirectory()) {
   auto repo = ITokenRepository::Build({.data_path = path});
   {
-    auto [token_list, err] = repo->List();
+    auto [token_list, err] = repo->GetTokenList();
 
     REQUIRE(err == Error::kOk);
     REQUIRE(token_list.empty());
@@ -23,8 +23,8 @@ static std::unique_ptr<ITokenRepository> MakeRepo(std::filesystem::path path = B
   permissions.mutable_read()->Add("bucket_1");
   permissions.mutable_write()->Add("bucket_2");
 
-  REQUIRE(repo->Create("token-1", permissions) == Error::kOk);
-  REQUIRE(repo->Create("token-2", permissions) == Error::kOk);
+  REQUIRE(repo->CreateToken("token-1", permissions) == Error::kOk);
+  REQUIRE(repo->CreateToken("token-2", permissions) == Error::kOk);
   return repo;
 }
 
@@ -37,7 +37,7 @@ TEST_CASE("auth::TokenRepository should create a token") {
   permissions.mutable_read()->Add("bucket_1");
   permissions.mutable_write()->Add("bucket_2");
 
-  auto [token, err] = repo->Create("token", std::move(permissions));
+  auto [token, err] = repo->CreateToken("token", std::move(permissions));
   REQUIRE(err == Error::kOk);
 
   SECTION("check value") {
@@ -46,31 +46,66 @@ TEST_CASE("auth::TokenRepository should create a token") {
   }
 
   SECTION("check if it's added") {
-    auto [token_list, _] = repo->List();
+    auto [token_list, _] = repo->GetTokenList();
     REQUIRE(token_list.size() == 1);
     REQUIRE(token_list[0].name() == "token");
   }
 
   SECTION("name must be unique") {
-    REQUIRE(repo->Create("token", {}).error == Error::Conflict("Token 'token' already exists"));
+    REQUIRE(repo->CreateToken("token", {}).error == Error::Conflict("Token 'token' already exists"));
   }
 
   SECTION("name can't be empty") {
-    REQUIRE(repo->Create("", {}).error == Error::UnprocessableEntity("Token name can't be empty"));
+    REQUIRE(repo->CreateToken("", {}).error == Error::UnprocessableEntity("Token name can't be empty"));
   }
 
   SECTION("must be persistent") {
     auto new_repo = ITokenRepository::Build({.data_path = path});
-    auto [token_list, _] = new_repo->List();
+    auto [token_list, _] = new_repo->GetTokenList();
     REQUIRE(token_list.size() == 1);
     REQUIRE(token_list[0].name() == "token");
+  }
+}
+
+TEST_CASE("auth::TokenRepository should update token") {
+  const auto path = BuildTmpDirectory();
+  auto repo = ITokenRepository::Build({.data_path = path});
+
+  ITokenRepository::TokenPermissions permissions;
+  permissions.set_full_access(false);
+  permissions.mutable_read()->Add("bucket_1");
+  permissions.mutable_write()->Add("bucket_2");
+
+  {
+    auto [_, err] = repo->CreateToken("new-token", std::move(permissions));
+    REQUIRE(err == Error::kOk);
+
+    ITokenRepository::TokenPermissions new_permissions;
+    new_permissions.set_full_access(true);
+
+    REQUIRE(repo->UpdateToken("new-token", new_permissions) == Error::kOk);
+  }
+
+  SECTION("check if it's updated") {
+    auto [token, _] = repo->FindByName("new-token");
+    REQUIRE(token.permissions().full_access());
+  }
+
+  SECTION("must be persistent") {
+    auto new_repo = ITokenRepository::Build({.data_path = path});
+    auto [token, _] = new_repo->FindByName("new-token");
+    REQUIRE(token.permissions().full_access());
+  }
+
+  SECTION("can't update non-existing token") {
+    REQUIRE(repo->UpdateToken("non-existing-token", {}) == Error::NotFound("Token 'non-existing-token' doesn't exist"));
   }
 }
 
 TEST_CASE("auth::TokenRepository should list tokens") {
   auto repo = MakeRepo();
 
-  auto [token_list, err] = repo->List();
+  auto [token_list, err] = repo->GetTokenList();
   REQUIRE(err == Error::kOk);
   REQUIRE(token_list.size() == 2);
 
@@ -108,7 +143,7 @@ TEST_CASE("auth::TokenRepository should find s token by value") {
   permissions.set_full_access(false);
   permissions.mutable_read()->Add("bucket_1");
 
-  auto [token_resp, err] = repo->Create("token-3", permissions);
+  auto [token_resp, err] = repo->CreateToken("token-3", permissions);
   REQUIRE(err == Error::kOk);
 
   auto [token, find_err] = repo->ValidateToken(token_resp.value());
@@ -127,11 +162,11 @@ TEST_CASE("auth::TokenRepository should remove token by name") {
   const auto path = BuildTmpDirectory();
   auto repo = MakeRepo(path);
 
-  REQUIRE(repo->Remove("token-1") == Error::kOk);
+  REQUIRE(repo->RemoveToken("token-1") == Error::kOk);
   REQUIRE(repo->FindByName("token-1").error.code == Error::kNotFound);
 
   SECTION("404 error") {
-    REQUIRE(repo->Remove("token-XXX") == Error::NotFound("Token 'token-XXX' doesn't exist"));
+    REQUIRE(repo->RemoveToken("token-XXX") == Error::NotFound("Token 'token-XXX' doesn't exist"));
   }
 
   SECTION("should be persistent") {